Refactor SoundSInk Class.
Just refeactoring
