+++ src/model/GameEnvironment.java
@@ -152,8 +152,8 @@ public class GameEnvironment {
         // Update sound sinks for all units
         for (BaseUnit unit : units) {
             // Notice that each unit is a source for its sink as well. If it is too noisy, it might not be able to perceive the surrounding.
-            unit.updateSoundSink(soundSources, terrain, surfaces, units);
-            unit.updatePerceivedSoundSink();
+            unit.getSoundSink().updateSoundSink(soundSources, terrain, surfaces, units, unit);
+            unit.getSoundSink().updatePerceivedSoundSink();
         }
 
         // Broadcast running, marching and arrow fire event events
